# backend/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Covara One API",
    "description": "Backend API for the Covara One parametric income-protection platform. Includes zero-touch auto-claims, live trigger ingestion (OpenWeather/TomTom/CPCB), 5-layer fraud detection, IRDAI-compliant KYC (Sandbox.co.in), and Twilio OTP + WhatsApp notifications.",
    "version": "0.5.0"
  },
  "paths": {
...
    "/payments/checkout-session": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create Stripe Checkout session for weekly plan purchase",
        "operationId": "create_checkout_session_payments_checkout_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/payments/checkout-session/{session_id}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Get Stripe Checkout session status",
        "operationId": "get_checkout_session_status_payments_checkout_session__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "auto_finalize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Auto Finalize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/payments/checkout/finalize": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Finalize paid Stripe checkout and activate policy",
        "operationId": "finalize_checkout_payments_checkout_finalize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizeCheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutSessionRequest": {
        "properties": {
          "plan": {
            "type": "string",
            "title": "Plan",
            "default": "essential"
          }
        },
        "type": "object",
        "title": "CheckoutSessionRequest"
      },
      "FinalizeCheckoutRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "FinalizeCheckoutRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}